requirements.txt
awscli==1.33.18
pytest==8.2.2
minio==7.2.7
s3fs==0.4.2
pandas==2.2.2
tables==3.10.1
pyarrow==17.0.0
scikit-learn==1.5.1
ray==2.34.0
dask==2024.8.1
distributed==2024.8.1
mosaicml-streaming==0.8.0
